refactor: extract test utils & login mixin; isort
